requirements.txt
ffmpeg-python 
openai-whisper 
pydub
transformers 
datasets
torchaudio
punctuators
stable-ts
pytube
huggingface_hub